gh-111968: Use per-thread freelists for float in free-threading
