+++ Python/pylifecycle.c
@@ -1757,10 +1757,10 @@ finalize_interp_types(PyInterpreterState *interp)
     _PySlice_Fini(interp);
 
     _PyUnicode_Fini(interp);
-    _PyFloat_Fini(interp);
 
     _PyFreeListState *state = _PyFreeListState_GET();
     _PyList_Fini(state);
+    _PyFloat_Fini(state);
 
 #ifdef Py_DEBUG
     _PyStaticObjects_CheckRefcnt(interp);
