+++ Python/pystate.c
@@ -1458,6 +1458,7 @@ clear_datastack(PyThreadState *tstate)
 void
 _Py_ClearFreeLists(_PyFreeListState *state, int is_finalization)
 {
+    _PyFloat_ClearFreeList(state, is_finalization);
     _PyList_ClearFreeList(state, is_finalization);
 }
